+++ scripts/start_xray_on_boot.sh
@@ -21,3 +21,4 @@ echo $! > "$PIDFILE"
 echo "Xray started with PID $(cat $PIDFILE)"
 
 
+